refactored elements and improved documentation of components
