+++ component/tile/reclassify/__init__.py
@@ -1,1 +1,2 @@
 from .reclassify import *
+from .datatable import *
